Work: Moving the Animation culling subsubsection
It now sits its own subsection at the beginning of the work section. This
allows all subsequent discussion in the work section to contain code snippets
with animation culling code. This will make creating the snippets easier. As
a result, the RNA optimizations are now also in their own subsection and the
Other Optimizations subsection has been removed.